+++ include/coli/utility.h
@@ -15,6 +15,7 @@
 
 #include <memory>
 #include <vector>
+#include <variant>
 #include <concepts>
 #include <algorithm>
 #include <typeindex>
@@ -35,6 +36,10 @@
 #include <glm/trigonometric.hpp>
 #include <glm/gtx/quaternion.hpp>
 #include <glm/gtc/matrix_transform.hpp>
+
+#include <bullet/btBulletDynamicsCommon.h>
+#include <bullet/btBulletCollisionCommon.h>
+#include <bullet/BulletCollision/CollisionShapes/btBox2dShape.h>
 
 /**
  * @brief For internal types.
@@ -149,6 +154,46 @@ namespace Coli::Utility
     extern template class COLI_EXPORT Hash<Types::rotator_type<true>>;
     extern template class COLI_EXPORT Hash<Types::rotator_type<false>>;
 #endif
+
+    class COLI_EXPORT LockFreeBase
+    {
+    protected:
+        LockFreeBase() noexcept;
+
+    public:
+        LockFreeBase(LockFreeBase const&) noexcept;
+        LockFreeBase(LockFreeBase&&) noexcept;
+
+        LockFreeBase& operator=(LockFreeBase const&) noexcept;
+        LockFreeBase& operator=(LockFreeBase&&) noexcept;
+
+    private:
+        std::atomic_flag myFlag = ATOMIC_FLAG_INIT;
+
+        friend class LockFreeTaker;
+    };
+
+    class COLI_EXPORT LockFreeTaker final
+    {
+    public:
+        LockFreeTaker() = delete;
+
+        LockFreeTaker(LockFreeBase& object) noexcept;
+
+        LockFreeTaker(LockFreeTaker&&) = delete;
+        LockFreeTaker(const LockFreeTaker&) = delete;
+
+        LockFreeTaker& operator=(LockFreeTaker&&) = delete;
+        LockFreeTaker& operator=(const LockFreeTaker&) = delete;
+
+        ~LockFreeTaker() noexcept;
+
+        [[nodiscard]] operator bool() const noexcept;
+
+    private:
+        LockFreeBase& myObject;
+        bool myFlag;
+    };
 }
 
 #endif