+++ tests/src/game/scene.cpp
@@ -45,6 +45,23 @@ TEST_F(SceneTest, CreateSuccess) {
     EXPECT_NO_THROW(Game::Scene scene);
 }
 
+/* Validity */
+
+TEST_F(SceneTest, CheckValidity)
+{
+    CREATE_SCENE
+    EXPECT_TRUE(scene->is_valid());
+}
+
+TEST_F(SceneTest, Reset)
+{
+    CREATE_SCENE
+    ASSERT_TRUE(scene->is_valid());
+
+    scene->reset();
+    EXPECT_FALSE(scene->is_valid());
+}
+
 /* Objects */
 
 TEST_F(SceneTest, AddObject)
@@ -63,59 +80,4 @@ TEST_F(SceneTest, RemoveObject)
 
     ASSERT_NO_THROW(object = scene->create());
     EXPECT_FALSE(object.expired());
-}
-
-TEST_F(SceneTest, Order)
-{
-    CREATE_SCENE;
-
-    for (int i = 0; i < 10; ++i)
-    {
-        auto object = scene->create();
-
-        if (i % 2)
-            object.emplace<Game::Components::Layer>(i * 50 + 2);
-    }
-
-    auto const ordered = scene->ordered();
-    auto comparator = [] (Game::ObjectHandle const& lhs, Game::ObjectHandle const& rhs)
-    {
-        constexpr size_t no_layer = std::numeric_limits<long long>::max();
-
-        auto const left = lhs.try_get<Game::Components::Layer>();
-        auto const right = rhs.try_get<Game::Components::Layer>();
-
-        auto const leftValue = left ? left->layer() : no_layer;
-        auto const rightValue = right ? right->layer() : no_layer;
-
-        return leftValue < rightValue;
-    };
-
-    EXPECT_TRUE(std::is_sorted(ordered.begin(), ordered.end(), comparator));
-}
-
-TEST_F(SceneTest, Filter)
-{
-    CREATE_SCENE;
-
-    size_t layered = 0;
-
-    for (int i = 0; i < 10; ++i)
-    {
-        auto object = scene->create();
-
-        if (i % 4) {
-            object.emplace<Game::Components::Layer>(i * 50 + 2);
-            ++layered;
-        }
-    }
-
-    auto const filtered = scene->filtered<Game::Components::Layer>();
-    auto predicate = [] (Game::ObjectHandle const& lhs) {
-        return lhs.contains<Game::Components::Layer>();
-    };
-
-    EXPECT_TRUE(
-        layered == filtered.size() &&
-        std::all_of(filtered.begin(), filtered.end(), predicate));
 }